refactoring: encapsulate logic in ItemDataRepoImpl.kt
